fix(embedding): address code review findings for MLX provider
Three blocking issues from code review:

1. RAM detection used syscall.Sysctl on hw.memsize (a uint64) and tried
   to parse it as a decimal string — always returned 0, defeating the
   <8GB guard. Fixed with unix.SysctlUint64.

2. Weight loader used f.Read for large blobs which can short-read.
   Switched to io.ReadFull for both the header and 160MB data blob.

3. MLX safetensors config shipped with empty ModelHash/TokenizerHash,
   skipping the downloader's integrity check. Populated with verified
   sha256 values.

Also fixed: softmax axis now uses explicit 3 instead of -1 for
consistency with layerNorm's negative-axis resolution; SplitAxis doc
comment corrected; EmbedWithPrefix now honors the prefix instead of
silently discarding it; memory profile test uses signed subtraction
to avoid uint64 wraparound.

+++ pkg/embedding/mlx_provider_test.go
@@ -123,7 +123,7 @@ func TestMLXProviderParity(t *testing.T) {
 		}
 
 		runtime.ReadMemStats(&m2)
-		allocAfter10Batches := m2.Alloc - m1.Alloc
+		allocAfter10Batches := int64(m2.Alloc) - int64(m1.Alloc)
 		t.Logf("Memory after 10 batches: alloc=%d bytes (%.1f MB), total_alloc=%d, sys=%d (%.1f MB)",
 			allocAfter10Batches,
 			float64(allocAfter10Batches)/1024/1024,
